Rename cond to match
