+++ composer.json
@@ -29,7 +29,6 @@
             "src/Functional/CompareOn.php",
             "src/Functional/CompareObjectHashOn.php",
             "src/Functional/Compose.php",
-            "src/Functional/Cond.php",
             "src/Functional/Contains.php",
             "src/Functional/Curry.php",
             "src/Functional/CurryN.php",
@@ -60,6 +59,7 @@
             "src/Functional/Last.php",
             "src/Functional/LastIndexOf.php",
             "src/Functional/Map.php",
+            "src/Functional/Match.php",
             "src/Functional/Maximum.php",
             "src/Functional/Memoize.php",
             "src/Functional/Minimum.php",
