+++ src/Functional/Match.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (C) 2011-2016 by Lars Strojny <lstrojny@php.net>
+ * Copyright (C) 2011-2016 by Gilles Crettenand <gilles@crettenand.info>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,26 +20,26 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-namespace Functional\Tests;
+namespace Functional;
 
-use function Functional\cond;
-use function Functional\equal;
-use function Functional\const_function;
+use Functional\const_function;
 
-class CondTest extends AbstractTestCase
+/**
+ * Performs an operation checking for the given conditions
+ *
+ * @param array $conditions the conditions to check against
+ *
+ * @return callable|null the function that calls the callable of the first truthy condition
+ */
+function match(array $conditions)
 {
-    public function testCond()
-    {
-        $test = cond([
-            [equal('foo'), const_function('is foo')],
-            [equal('bar'), const_function('is bar')],
-            [equal('baz'), const_function('is baz')],
-            [const_function(true), function ($x) { return 'default is '.$x; }],
-        ]);
+    return function ($x) use ($conditions) {
+        foreach ($conditions as $condition) {
+            if ($condition[0]($x)) {
+                return $condition[1]($x);
+            }
+        }
 
-        $this->assertEquals('is foo', $test('foo'));
-        $this->assertEquals('is bar', $test('bar'));
-        $this->assertEquals('is baz', $test('baz'));
-        $this->assertEquals('default is qux', $test('qux'));
-    }
+        return const_function(null);
+    };
 }
